+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.28
+
+* Add functionality to automatically scale table images so that the text height is optimum for
+  `tesseract` inference
+
 ## 0.5.27
 
 * table structure detection now pads the input image by 25 pixels in all 4 directions to improve its recall